requirements.txt
Django==2.0
pytz==2017.3
gunicorn==19.7.1
dj-database-url==0.5.0
whitenoise==3.3.1
psycopg2==2.7.5
djangorestframework==3.8.2
line-bot-sdk==1.7.2